backend/requirements.txt
Django==5.2.7

djangorestframework==3.16.1

django-cors-headers==4.9.0

django-filter==25.2

# Google Places API
googlemaps==4.10.0

# Environment variables
python-decouple==3.8

# HTTP requests
requests==2.31.0

# Optional: MongoDB support (uncomment if using Mongo)
# djongo==1.3.6
# pymongo==4.10.1
